Refactor package structure and update imports for tools and main entry point
