+++ py_code/mcp_server/__main__.py
@@ -1,0 +1,10 @@
+from py_code.integrated_server import run_server
+
+
+def main():
+    """Entry point for the package."""
+    run_server()
+
+
+if __name__ == "__main__":
+    main()
