--- py_code_mcp_server/__main__.py
@@ -1,4 +1,0 @@
-from py_code.integrated_server import run_server
-
-if __name__ == "__main__":
-    run_server()
